refactor: context, provider 한 파일로 합치고, sse, reset context는 베이스 context를 creteContext에서 만들고, 해당 provider들은 한번에 묶어서 export
