+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import { Router } from 'pages';
-import { AlertProvider, ApiProvider } from 'utils/provider';
+import { AlertProvider, ApiProvider } from 'utils';
 import 'i18n';
 import './index.css';
