+++ src/pages/pagesData.tsx
@@ -3,23 +3,17 @@ import { Home } from './home';
 import { Login } from './login';
 import { Signup } from './signup';
 import { FindPw } from './findPw';
-import {
-  CreateResetProvider,
-  SearchResetProvider,
-  DashboardResetProvider,
-} from 'utils';
+import { SSEProviders, ResetProviders } from 'utils';
 
 const pagesData: routerType[] = [
   {
     path: '*',
     element: (
-      <CreateResetProvider>
-        <SearchResetProvider>
-          <DashboardResetProvider>
-            <Home />
-          </DashboardResetProvider>
-        </SearchResetProvider>
-      </CreateResetProvider>
+      <SSEProviders>
+        <ResetProviders>
+          <Home />
+        </ResetProviders>
+      </SSEProviders>
     ),
     title: 'home',
   },
