--- src/utils/context/ApiContext.ts
@@ -1,15 +1,0 @@
-import { createContext } from 'react';
-
-type ApiContextType = {
-  authApi: any;
-  refreshableApi: any;
-  checkTokenFromCookie: () => void;
-};
-
-const ApiContext = createContext<ApiContextType>({
-  authApi: null,
-  refreshableApi: null,
-  checkTokenFromCookie: () => {},
-});
-
-export default ApiContext;
