+++ src/utils/contexts/TagContext.tsx
@@ -1,9 +1,42 @@
 import { Tag } from 'pages/home/subPages/interfaces';
+import { createContext, Dispatch, SetStateAction } from 'react';
 import { useState, useCallback, ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
-import { TagContext, TagModalState } from 'utils';
 
-const TagProvider = ({ children }: { children: ReactNode }) => {
+type TagModalState = {
+  mode: 'create' | 'edit';
+  isOpen: boolean;
+  tag: Tag | null;
+  inputTagName: string;
+  onClose: () => void;
+  title: string;
+};
+
+interface TagContextType {
+  selectedTag: Tag | null;
+  setSelectedTag: Dispatch<SetStateAction<Tag | null>>;
+  tagStack: Tag[];
+  setTagStack: Dispatch<SetStateAction<Tag[]>>;
+  tagModal: TagModalState | null;
+  openTagEditModal: (tag: Tag) => void;
+  openTagCreateModal: (parentTag: Tag | null) => void;
+  closeTagModal: () => void;
+  onTagReset: () => void;
+}
+
+export const TagContext = createContext<TagContextType>({
+  selectedTag: null,
+  setSelectedTag: (_: SetStateAction<Tag | null>) => {},
+  tagStack: [],
+  setTagStack: (_: SetStateAction<Tag[]>) => {},
+  tagModal: null,
+  openTagEditModal: (_: Tag) => {},
+  openTagCreateModal: (_: Tag | null) => {},
+  closeTagModal: () => {},
+  onTagReset: () => {},
+});
+
+export const TagProvider = ({ children }: { children: ReactNode }) => {
   const [tagModal, setTagModal] = useState<TagModalState | null>(null);
   const [selectedTag, setSelectedTag] = useState<Tag | null>(null);
   const [tagStack, setTagStack] = useState<Tag[]>([]);
@@ -59,5 +92,3 @@ const TagProvider = ({ children }: { children: ReactNode }) => {
     </TagContext.Provider>
   );
 };
-
-export default TagProvider;
