+++ src/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './context';
+export * from './contexts';
 export * from './dialog';
-export * from './provider';
